Create list.c
Queue messages of a given length, in which messages are sorted by priority. 
New messages must be added according to priority. 
If the queue is complete, then to add a new high priority message, 
the lowest priority message must be deleted. 
Use the list to organize the queue.

The lower the number, the higher the priority